+++ frontend/src/models/User/auth.ts
@@ -1,0 +1,8 @@
+export function checkUserAccess(): string {
+  const token = localStorage.getItem('token');
+  if (!token) {
+    window.location.href = '/login';
+    throw new Error('Token does not exist, redirected to login page.');
+  }
+  return token;
+}
